+++ packages/ckeditor5-dev-docs/CHANGELOG.md
@@ -1,6 +1,18 @@
 Changelog
 =========
 
+## [6.2.3](https://github.com/ckeditor/ckeditor5-dev/compare/@ckeditor/ckeditor5-dev-docs@6.2.2...@ckeditor/ckeditor5-dev-docs@6.2.3) (2017-05-18)
+
+### Bug fixes
+
+* Use more reliable way to resolve module paths. Closes [#206](https://github.com/ckeditor/ckeditor5-dev/issues/206). ([41e6a95](https://github.com/ckeditor/ckeditor5-dev/commit/41e6a95))
+
+
+## [6.2.2](https://github.com/ckeditor/ckeditor5-dev/compare/@ckeditor/ckeditor5-dev-docs@6.2.1...@ckeditor/ckeditor5-dev-docs@6.2.2)
+
+Internal changes only (updated dependencies, documentation, etc.).
+
+
 ## [6.2.1](https://github.com/ckeditor/ckeditor5-dev/compare/@ckeditor/ckeditor5-dev-docs@6.2.0...@ckeditor/ckeditor5-dev-docs@6.2.1) (2017-04-27)
 
 Internal changes only (updated dependencies, documentation, etc.).