+++ packages/ckeditor5-dev-tests/CHANGELOG.md
@@ -1,6 +1,24 @@
 Changelog
 =========
 
+## [7.4.4](https://github.com/ckeditor/ckeditor5-dev/compare/@ckeditor/ckeditor5-dev-tests@7.4.3...@ckeditor/ckeditor5-dev-tests@7.4.4) (2017-05-17)
+
+### Bug fixes
+
+* The `packages/` dir needs to be ignored when cloning subpackages on Travis. Closes [#203](https://github.com/ckeditor/ckeditor5-dev/issues/203). ([0b4080f](https://github.com/ckeditor/ckeditor5-dev/commit/0b4080f))
+
+
+## [7.4.3](https://github.com/ckeditor/ckeditor5-dev/compare/@ckeditor/ckeditor5-dev-tests@7.4.2...@ckeditor/ckeditor5-dev-tests@7.4.3) (2017-05-16)
+
+### Bug fixes
+
+* The test tool needs to install eslint-config-ckeditor5 on Travis. ([787418b](https://github.com/ckeditor/ckeditor5-dev/commit/787418b))
+
+## 7.4.2 (2017-05-16)
+
+Internal changes only (updated dependencies, documentation, etc.).
+
+
 ## [7.4.1](https://github.com/ckeditor/ckeditor5-dev/compare/@ckeditor/ckeditor5-dev-tests@7.4.0...@ckeditor/ckeditor5-dev-tests@7.4.1) (2017-04-27)
 
 Internal changes only (updated dependencies, documentation, etc.).